refactor(core): Extract Command Palette into command_palette.rs
Move all command-palette WindowManager methods (wm_menu_items,
open/close command palette, tab-outline mode, bounds/visibility/key,
event handling) from mod.rs, overlays.rs, and command_menu.rs into a
dedicated command_palette.rs submodule. Delete the now-empty
command_menu.rs.

Co-locate the palette's unit tests in command_palette.rs (relocated
TestOverlay to module scope) and add coverage for the state
transitions: input-mode toggling, overlay registration/removal,
tab-outline expiry propagation, and event-driven dismissal.
