--- crates/term-wm-core/src/window/window_manager/command_menu.rs
@@ -1,13 +1,0 @@
-use super::WindowManager;
-use crate::actions::TermWmAction;
-use crate::components::{Component, Overlay, WmComponent};
-
-impl<C: Component<TermWmAction>, L: WmComponent, O: Overlay<TermWmAction>> WindowManager<C, L, O> {
-    pub fn command_menu_visible(&self) -> bool {
-        self.command_palette_visible()
-    }
-
-    pub fn close_command_menu(&mut self) {
-        self.close_command_palette();
-    }
-}
