+++ crates/term-wm-core/src/window/window_manager/mod.rs
@@ -1,5 +1,5 @@
 mod chrome;
-mod command_menu;
+mod command_palette;
 mod drag;
 mod focus;
 pub(crate) mod layer_manager;
@@ -2323,38 +2323,8 @@ impl<C: Component<TermWmAction>, L: WmComponent, O: Overlay<TermWmAction>> Windo
         self.register_overlay::<system_tags::ExitConfirm>(key);
     }
 
-    pub fn open_command_palette_overlay(&mut self, overlay: O) {
-        let key = self.overlays.insert(overlay);
-        self.register_overlay::<system_tags::CommandPalette>(key);
-        self.input_mode = crate::actions::WmInputMode::CommandPalette;
-    }
-
     pub fn set_scroll_keyboard_enabled(&mut self, enabled: bool) {
         self.scroll_keyboard_enabled_default = enabled;
-    }
-
-    /// Enter tab outline mode — palette becomes dim overlay, panels hide in monocle.
-    pub fn set_tab_outline_mode(&mut self, duration: Duration) {
-        let expires = Instant::now() + duration;
-        self.tab_outline_until = Some(expires);
-        if let Some(key) = self.get_overlay::<system_tags::CommandPalette>()
-            && let Some(overlay) = self.overlays.get_mut(key)
-        {
-            overlay.set_tab_outline(Some(expires));
-        }
-        if let Some(handle) = &self.system_task_handle {
-            let _ = handle.schedule_once(duration, crate::actions::SystemTask::ClearTabOutline);
-        }
-    }
-
-    /// Clear tab outline mode — restore palette/panels to normal.
-    pub fn clear_tab_outline(&mut self) {
-        self.tab_outline_until = None;
-        if let Some(key) = self.get_overlay::<system_tags::CommandPalette>()
-            && let Some(overlay) = self.overlays.get_mut(key)
-        {
-            overlay.set_tab_outline(None);
-        }
     }
 
     /// Whether tab outline mode is currently active.
@@ -2675,201 +2645,6 @@ impl<C: Component<TermWmAction>, L: WmComponent, O: Overlay<TermWmAction>> Windo
         }
         btns
     }
-
-    pub fn wm_menu_items(
-        &self,
-    ) -> Vec<crate::components::MenuDisplayItem<crate::actions::TermWmAction>> {
-        use crate::components::{MenuDisplayItem, MenuItem};
-        use crate::window::WindowState;
-        use crate::window::window_manager::system_tags;
-
-        let debug_log_visible = self
-            .get_system_window::<system_tags::DebugLog>()
-            .is_some_and(|k| self.window_state(k) == Some(WindowState::Mapped));
-        let system_panel_visible = self
-            .get_system_window::<system_tags::SystemPanel>()
-            .is_some_and(|k| self.window_state(k) == Some(WindowState::Mapped));
-
-        let mouse_label = if self.mouse_capture_enabled {
-            "Mouse: Disable Capture"
-        } else {
-            "Mouse: Enable Capture"
-        };
-        let clipboard_label = if self.clipboard_enabled {
-            "Clipboard: Disable"
-        } else {
-            "Clipboard: Enable"
-        };
-        let selection_label = if self.window_selection_enabled {
-            "Clipboard: Disable Selection"
-        } else {
-            "Clipboard: Enable Selection"
-        };
-        let debug_label = if debug_log_visible {
-            "System: Disable Debug Log"
-        } else {
-            "System: Enable Debug Log"
-        };
-        let panel_label = if system_panel_visible {
-            "System: Disable Panel"
-        } else {
-            "System: Enable Panel"
-        };
-
-        fn mi(
-            label: &'static str,
-            icon: Option<&'static str>,
-            action: crate::actions::TermWmAction,
-        ) -> MenuDisplayItem<crate::actions::TermWmAction> {
-            MenuDisplayItem::Item(MenuItem {
-                label: label.into(),
-                icon,
-                action,
-                disabled: false,
-            })
-        }
-
-        let focused = self.focused_window();
-        let has_active = self.windows.contains_key(focused);
-
-        let mut items: Vec<MenuDisplayItem<crate::actions::TermWmAction>> = vec![
-            // Top group
-            mi("Resume", Some("▶"), crate::actions::TermWmAction::CloseMenu),
-            mi(
-                "New Window",
-                Some("+"),
-                crate::actions::TermWmAction::NewWindow,
-            ),
-            MenuDisplayItem::Separator,
-        ];
-
-        // Window management group (directly below top group)
-        {
-            if has_active {
-                let raw_title = self.window_title(focused);
-                let title = crate::utils::truncate_with_ellipsis(&raw_title, 25);
-                let super_key = self
-                    .keybindings()
-                    .combos_for(crate::actions::TermWmAction::OpenCommandPalette)
-                    .first()
-                    .map(|c| c.to_string())
-                    .unwrap_or_else(|| "Super".to_string());
-
-                // Send SUPER key to window
-                items.push(MenuDisplayItem::Item(MenuItem {
-                    label: format!("Send {} to {}", super_key, title).into(),
-                    icon: Some("A"),
-                    action: crate::actions::TermWmAction::SendSuperKeyToWindow(focused),
-                    disabled: false,
-                }));
-
-                // Close window
-                items.push(MenuDisplayItem::Item(MenuItem {
-                    label: format!("Close {}", title).into(),
-                    icon: Some("X"),
-                    action: crate::actions::TermWmAction::CloseWindow(focused),
-                    disabled: false,
-                }));
-
-                // Maximize / Restore
-                let is_maxed = self.window(focused).is_some_and(|w| w.is_maximized());
-                if !self.is_monocle() {
-                    items.push(MenuDisplayItem::Item(MenuItem {
-                        label: (if is_maxed {
-                            format!("Restore {}", title)
-                        } else {
-                            format!("Maximize {}", title)
-                        })
-                        .into(),
-                        icon: Some(if is_maxed { "─" } else { "▢" }),
-                        action: crate::actions::TermWmAction::MaximizeWindow(focused),
-                        disabled: false,
-                    }));
-                    items.push(MenuDisplayItem::Item(MenuItem {
-                        label: format!("Minimize {}", title).into(),
-                        icon: Some("_"),
-                        action: crate::actions::TermWmAction::MinimizeWindow(focused),
-                        disabled: false,
-                    }));
-                }
-
-                // Switch to windows
-                for (key, switch_title) in self.window_titles() {
-                    items.push(MenuDisplayItem::Item(MenuItem {
-                        label: format!("Switch to: {}", switch_title).into(),
-                        icon: Some("→"),
-                        action: crate::actions::TermWmAction::FocusWindow(key),
-                        disabled: key == focused,
-                    }));
-                }
-            }
-        }
-
-        // View group
-        {
-            items.push(MenuDisplayItem::Separator);
-            items.push(mi(
-                self.monocle_mode.action_label(),
-                Some("▢"),
-                crate::actions::TermWmAction::ToggleMonocle,
-            ));
-            {
-                let label = if self.managed_layout.is_some() {
-                    "View: Float Windows"
-                } else {
-                    "View: Tile Windows"
-                };
-                let mut item = mi(label, Some("⊞"), crate::actions::TermWmAction::ToggleTiling);
-                if self.is_monocle()
-                    && let MenuDisplayItem::Item(ref mut mi) = item
-                {
-                    mi.disabled = true;
-                }
-                items.push(item);
-            }
-        }
-
-        // Settings group
-        {
-            items.push(MenuDisplayItem::Separator);
-            items.push(mi(
-                mouse_label,
-                Some("◆"),
-                crate::actions::TermWmAction::ToggleMouseCapture,
-            ));
-            items.push(mi(
-                clipboard_label,
-                Some("■"),
-                crate::actions::TermWmAction::ToggleClipboardMode,
-            ));
-            items.push(mi(
-                selection_label,
-                Some("●"),
-                crate::actions::TermWmAction::ToggleWindowSelection,
-            ));
-            items.push(mi(
-                debug_label,
-                Some("≣"),
-                crate::actions::TermWmAction::ToggleDebugWindow,
-            ));
-            items.push(mi(
-                panel_label,
-                Some("*"),
-                crate::actions::TermWmAction::ToggleSystemPanel,
-            ));
-
-            // Help/Exit as last group
-            items.push(MenuDisplayItem::Separator);
-            items.push(mi("Help", Some("?"), crate::actions::TermWmAction::Help));
-            items.push(mi(
-                "Exit UI",
-                Some("⏻"),
-                crate::actions::TermWmAction::ExitUi,
-            ));
-        }
-
-        items
-    }
 }
 
 #[derive(Clone)]
@@ -2963,6 +2738,40 @@ fn make_keys<L: WmComponent, O: Overlay<TermWmAction>>(
     (0..n)
         .map(|_| wm.create_window(TestComponent::Noop(crate::components::NoopComponent)))
         .collect()
+}
+
+#[cfg(test)]
+struct TestOverlay {
+    bounds: Option<LayoutRect>,
+}
+
+#[cfg(test)]
+impl Component<TermWmAction> for TestOverlay {
+    fn render(
+        &mut self,
+        _backend: &mut dyn term_wm_render::RenderBackend,
+        _area: LayoutRect,
+        _ctx: &crate::component_context::ComponentContext,
+        _registry: &mut crate::hitbox_registry::HitboxRegistry,
+    ) {
+    }
+    fn update(
+        &mut self,
+        _action: TermWmAction,
+        _ctx: &crate::component_context::ComponentContext,
+        _actions: &mut VecDeque<(WindowKey, TermWmAction)>,
+    ) {
+    }
+}
+
+#[cfg(test)]
+impl Overlay<TermWmAction> for TestOverlay {
+    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
+        self
+    }
+    fn render_area(&self) -> Option<LayoutRect> {
+        self.bounds.filter(|b| b.width > 0 && b.height > 0)
+    }
 }
 
 #[cfg(test)]
@@ -6857,38 +6666,6 @@ mod tests {
     }
 
     #[test]
-    fn command_palette_empty_map_returns_no_action() {
-        let mut wm = WindowManager::<TestComponent>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        assert!(!wm.command_menu_visible());
-        let event = crate::events::Event::Key(crate::events::KeyEvent {
-            code: crate::events::KeyCode::Esc,
-            modifiers: crate::events::KeyModifiers::NONE,
-            kind: crate::events::KeyKind::Press,
-        });
-        assert!(wm.handle_command_palette_event(&event).is_none());
-    }
-
-    #[test]
-    fn command_menu_visible_derived_from_overlay_map() {
-        let mut wm = WindowManager::<TestComponent>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        assert!(!wm.command_menu_visible());
-        wm.close_command_menu();
-        assert!(!wm.command_menu_visible());
-    }
-
-    #[test]
     fn tile_window_key_respects_quadrant() {
         use crate::layout::{Direction, LayoutNode, TilingLayout};
         use crate::window::{FloatRect, FloatRectSpec};
@@ -7006,114 +6783,6 @@ mod tests {
         // ActionRecorder returns None by default
         let result = wm.take_alternate_screen_transition(keys[0]);
         assert_eq!(result, None);
-    }
-
-    // ── TestOverlay for command_palette_bounds tests ──────────────────────
-
-    struct TestOverlay {
-        bounds: Option<LayoutRect>,
-    }
-
-    impl Component<TermWmAction> for TestOverlay {
-        fn render(
-            &mut self,
-            _backend: &mut dyn term_wm_render::RenderBackend,
-            _area: LayoutRect,
-            _ctx: &crate::component_context::ComponentContext,
-            _registry: &mut crate::hitbox_registry::HitboxRegistry,
-        ) {
-        }
-        fn update(
-            &mut self,
-            _action: TermWmAction,
-            _ctx: &crate::component_context::ComponentContext,
-            _actions: &mut VecDeque<(WindowKey, TermWmAction)>,
-        ) {
-        }
-    }
-
-    impl Overlay<TermWmAction> for TestOverlay {
-        fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-            self
-        }
-        fn render_area(&self) -> Option<LayoutRect> {
-            self.bounds.filter(|b| b.width > 0 && b.height > 0)
-        }
-    }
-
-    // ── command_palette_bounds tests ──────────────────────────────────────
-
-    #[test]
-    fn command_palette_bounds_returns_none_when_no_key() {
-        let wm = WindowManager::<TestComponent, NoopWmComponent, TestOverlay>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        assert_eq!(wm.command_palette_bounds(), None);
-    }
-
-    #[test]
-    fn command_palette_bounds_delegates_to_overlay_render_area() {
-        let mut wm = WindowManager::<TestComponent, NoopWmComponent, TestOverlay>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        let expected = LayoutRect {
-            x: 10,
-            y: 5,
-            width: 40,
-            height: 10,
-        };
-        let overlay = TestOverlay {
-            bounds: Some(expected),
-        };
-        wm.open_command_palette_overlay(overlay);
-        assert_eq!(wm.command_palette_bounds(), Some(expected));
-    }
-
-    #[test]
-    fn command_palette_bounds_returns_none_for_zero_size() {
-        let mut wm = WindowManager::<TestComponent, NoopWmComponent, TestOverlay>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        let overlay = TestOverlay {
-            bounds: Some(LayoutRect::default()),
-        };
-        wm.open_command_palette_overlay(overlay);
-        // zero-size is filtered by render_area() → None
-        assert_eq!(wm.command_palette_bounds(), None);
-    }
-
-    #[test]
-    fn command_palette_bounds_returns_none_for_missing_overlay() {
-        let mut wm = WindowManager::<TestComponent, NoopWmComponent, TestOverlay>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        // Insert a different overlay but don't set command_palette_key
-        let other = TestOverlay {
-            bounds: Some(LayoutRect {
-                x: 0,
-                y: 0,
-                width: 10,
-                height: 10,
-            }),
-        };
-        wm.overlays.insert(other);
-        assert_eq!(wm.command_palette_bounds(), None);
     }
 
     // ── handle_outside_click tests ───────────────────────────────────────
@@ -7365,18 +7034,6 @@ mod tests {
             std::collections::HashMap::new(),
         );
         assert_eq!(wm.help_key(), None);
-    }
-
-    #[test]
-    fn command_palette_key_returns_none_initially() {
-        let wm = WindowManager::<TestComponent>::with_config(
-            WmConfig::standalone(),
-            Arc::new(AppContext::new("test", "0.0.0")),
-            None,
-            crate::window::LayerManager::new(),
-            std::collections::HashMap::new(),
-        );
-        assert_eq!(wm.command_palette_key(), None);
     }
 
     // ── help_overlay_bounds tests ──────────────────────────────────────────
